feat: 采集配置导入功能 # Reviewed, transaction id: 17094
