+++ bklog/web/src/components/collection-access/components/index-import-modal/index.scss
@@ -1,0 +1,7 @@
+.index-import-modal {
+  .top-sync-select {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+  } 
+}
